Fix multiple-edge display
This patch addresses an accidental collapse of multiple edges between
nodes.  The effect is most promininent in `ProvenanceRecord`s that
extend the history of an object (such as an exhibit transfer) - it
should have both the `hadMember` and `wasDerivedFrom` relationships with
that object.  Prior to this patch, only one edge between node pairs
would be recorded, so one of those edges ended up hidden.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>