Speed Communication SPARQL queries
The Communication and wasInformedBy queries were written to account for
actions that generated a non-`ProvenanceRecord` object.  This was
observed to create a significant slowdown in a trial using generated
data (somewhat more than is practical to write by hand, unlike the data
used to test this query in this repository).

Review of the query showed the generating action actually had no impact
on results, so it was trimmed.

The inner subquery was further optimized to make the most
dependended-upon loop be only the `InvestigativeAction`s linked by a
`ProvenanceRecord`.

A reference was found to prevent generating extraneous blank nodes per
hidden binding of the non-`ProvenanceRecord` object.

No effect was observed on generated Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>