Port wasDerivedFrom CASE usage to PROV
This covers the use case where CASE data includes `wasDerivedFrom`
statements, and the PROV mapping can be used to confirm the
`wasDerivedFrom` statements in CASE are as well supported as the
`wasDerivedFrom` statements constructed from other `case-prov` mapping
queries.

From the current design vocabulary, if a CASE `wasDerivedFrom` statement
is in the graph, but independently asserted without support from the
chain of custody, that `wasDerivedFrom` edge would display as
"unqualified" (a dashed line).

As a reminder, CASE's `wasDerivedFrom` was defined in CASE 0.4.0 to act
like PROV-O's `wasDerivedFrom`, but without entailing domain or range
inferences.

No effects on Make-managed files were observed.

References:
* [ONT-390] (CP-21) Add transitive property to report causal chain of a
  derived object
* https://caseontology.org/releases/0.4.0/

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>