Add inferred classes in CASE property mappings
An example in CASE-Examples uses `case-investigation:wasDerivedFrom`
without setting up a provenance chain.  This triggered a missing-entity
error in `case_prov_dot.py`.  (The example is being added to review in
another PR in progress.)

When `wasDerivedFrom` is mapped from an explicit statement in CASE to
PROV-O, PROV-O class assignments should be "inferred" (generated) for
the triple's subject and object.

No effects were observed on Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>